debugged double hist issue?

+++ src/models/improvements.py
@@ -11,7 +11,7 @@ class PrefixSumCounts(nn.Module):
         self.d_vocab = d_vocab
 
     def forward(self, x):
-        one_hot = F.one_hot(x, num_classes=self.d_vocab).float()
+        one_hot = F.one_hot(x.long(), num_classes=self.d_vocab).float()
         cumsum = torch.cumsum(one_hot, dim=1)
         idx = x.unsqueeze(-1)
         counts = torch.gather(cumsum, dim=2, index=idx).float()
